+++ src/features/residents/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export * from './data/mockResidents';
 export { default as ResidentsPage } from './pages/ResidentsPage';
 export { default as ResidentProfilePage } from './pages/ResidentProfilePage';
+export { default as ResidentOnboardingWizard } from './components/ResidentOnboardingWizard';